+++ CHANGELOG.md
@@ -12,10 +12,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for CatchmentCNCLM51.
+- Added ntasks-per-node 
 
 ### Changed
 
+- Cleaned up ldas_setup. Split it to ldas.py and setup_utils.py, 
+
 ### Fixed
+
+- Fixed Restart = 1 when the domain is not global
 
 ### Removed